Flatten more screens
