+++ apps/app/src/components/ProductContext.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { useProductQuery, useRelatedProductsQuery } from '../../data/queries';
+import { useProductQuery, useRelatedProductsQuery } from '../data/queries';
 import {
 	useAddToCartMutation,
 	useUpdateCartProductMutation
-} from '../../data/mutations';
-import type { Product, ProductViewerContext } from '../../data/types';
+} from '../data/mutations';
+import type { Product, ProductViewerContext } from '../data/types';
 import { RouteProp, useNavigation, useRoute } from '@react-navigation/native';
-import { AppStackParamList } from '../../navigation/types';
+import { AppStackParamList } from '../navigation/types';
 import { View } from 'react-native';
 
 interface ProductContextType {
