+++ apps/app/src/components/home/FollowedStores.tsx
@@ -19,6 +19,13 @@ interface FollowedStoresProps {
 const FollowedStores: React.FC<FollowedStoresProps> = ({ followed }) => {
 	const { navigate } = useNavigation<NavigationProp<HomeStackParamList>>();
 
+	const handleStorePress = React.useCallback(
+		(storeId: string) => () => {
+			navigate('Home.Store', { storeId });
+		},
+		[]
+	);
+
 	return followed.length === 0 ? (
 		<Dialog
 			style={styles.dialog}
@@ -35,38 +42,22 @@ const FollowedStores: React.FC<FollowedStoresProps> = ({ followed }) => {
 					onPress: () => navigate('Home.FollowedStores')
 				}}
 			/>
+
 			<Spacer y={8} />
-			<FollowedStoresMain followed={followed} />
+
+			<FlatList
+				horizontal
+				data={followed.slice(0, 8)}
+				keyExtractor={item => item.id}
+				contentContainerStyle={{ paddingLeft: 16, gap: 12 }}
+				renderItem={({ item }) => (
+					<FollowedStoresItem
+						store={item}
+						onPress={handleStorePress(item.id)}
+					/>
+				)}
+			/>
 		</View>
-	);
-};
-
-interface FollowedStoresMainProps {
-	followed: Store[];
-}
-
-const FollowedStoresMain: React.FC<FollowedStoresMainProps> = ({
-	followed
-}) => {
-	const { navigate } = useNavigation<NavigationProp<HomeStackParamList>>();
-
-	const handleStorePress = React.useCallback(
-		(storeId: string) => () => {
-			navigate('Home.Store', { storeId });
-		},
-		[]
-	);
-
-	return (
-		<FlatList
-			horizontal
-			data={followed.slice(0, 8)}
-			keyExtractor={item => item.id}
-			contentContainerStyle={{ paddingLeft: 16, gap: 12 }}
-			renderItem={({ item }) => (
-				<FollowedStoresItem store={item} onPress={handleStorePress(item.id)} />
-			)}
-		/>
 	);
 };
