+++ apps/app/src/components/home/RecentOrders.tsx
@@ -101,4 +101,5 @@ const styles = StyleSheet.create({
 		marginLeft: 8
 	}
 });
+
 export default RecentOrders;
